+++ examples/example/lib/nested.auto_mappr.dart
@@ -81,20 +81,14 @@ class $Mappr implements _i1.AutoMapprInterface {
   TARGET? tryConvert<SOURCE, TARGET>(
     SOURCE? model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
-      return _safeConvert(
-        model,
-        onMappingError: onMappingError,
-      );
+      return _safeConvert(model, onMappingError: onMappingError);
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvert(
-          model,
-          onMappingError: onMappingError,
-        );
+        return mappr.tryConvert(model, onMappingError: onMappingError);
       }
     }
 
@@ -126,18 +120,16 @@ class $Mappr implements _i1.AutoMapprInterface {
   Iterable<TARGET?> tryConvertIterable<SOURCE, TARGET>(
     Iterable<SOURCE?> model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
       return model.map<TARGET?>(
-          (item) => _safeConvert(item, onMappingError: onMappingError));
+        (item) => _safeConvert(item, onMappingError: onMappingError),
+      );
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertIterable(
-          model,
-          onMappingError: onMappingError,
-        );
+        return mappr.tryConvertIterable(model, onMappingError: onMappingError);
       }
     }
 
@@ -169,7 +161,7 @@ class $Mappr implements _i1.AutoMapprInterface {
   List<TARGET?> tryConvertList<SOURCE, TARGET>(
     Iterable<SOURCE?> model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
       return tryConvertIterable<SOURCE, TARGET>(
@@ -179,10 +171,7 @@ class $Mappr implements _i1.AutoMapprInterface {
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertList(
-          model,
-          onMappingError: onMappingError,
-        );
+        return mappr.tryConvertList(model, onMappingError: onMappingError);
       }
     }
 
@@ -214,7 +203,7 @@ class $Mappr implements _i1.AutoMapprInterface {
   Set<TARGET?> tryConvertSet<SOURCE, TARGET>(
     Iterable<SOURCE?> model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
       return tryConvertIterable<SOURCE, TARGET>(
@@ -224,10 +213,7 @@ class $Mappr implements _i1.AutoMapprInterface {
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertSet(
-          model,
-          onMappingError: onMappingError,
-        );
+        return mappr.tryConvertSet(model, onMappingError: onMappingError);
       }
     }
 
@@ -267,7 +253,9 @@ class $Mappr implements _i1.AutoMapprInterface {
         return null;
       }
       return (_map__i2$NestedTagDto_To__i2$NestedTag(
-          (model as _i2.NestedTagDto?)) as TARGET);
+            (model as _i2.NestedTagDto?),
+          )
+          as TARGET);
     }
     throw Exception('No ${model.runtimeType} -> $targetTypeOf mapping.');
   }
@@ -275,19 +263,13 @@ class $Mappr implements _i1.AutoMapprInterface {
   TARGET? _safeConvert<SOURCE, TARGET>(
     SOURCE? model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (!useSafeMapping<SOURCE, TARGET>()) {
-      return _convert(
-        model,
-        canReturnNull: true,
-      );
+      return _convert(model, canReturnNull: true);
     }
     try {
-      return _convert(
-        model,
-        canReturnNull: true,
-      );
+      return _convert(model, canReturnNull: true);
     } catch (e, s) {
       onMappingError?.call(e, s, model);
       return null;
@@ -305,8 +287,9 @@ class $Mappr implements _i1.AutoMapprInterface {
     final model = input;
     if (model == null) {
       throw Exception(
-          r'Mapping UserDto → User failed because UserDto was null, and no default value was provided. '
-          r'Consider setting the whenSourceIsNull parameter on the MapType<UserDto, User> to handle null values during mapping.');
+        r'Mapping UserDto → User failed because UserDto was null, and no default value was provided. '
+        r'Consider setting the whenSourceIsNull parameter on the MapType<UserDto, User> to handle null values during mapping.',
+      );
     }
     return _i2.User(
       id: model.id,
@@ -314,9 +297,11 @@ class $Mappr implements _i1.AutoMapprInterface {
       nestedItems: model.nestedItems
           .map<_i2.Nested>((value) => _map__i2$NestedDto_To__i2$Nested(value))
           .toList(),
-      nestedItemsNullable: model.nestedItemsNullable
+      nestedItemsNullable:
+          model.nestedItemsNullable
               ?.map<_i2.Nested>(
-                  (value) => _map__i2$NestedDto_To__i2$Nested(value))
+                (value) => _map__i2$NestedDto_To__i2$Nested(value),
+              )
               .toList() ??
           <_i2.Nested>[],
       nestedItemsNullable2: model.nestedItemsNullable2
@@ -328,7 +313,8 @@ class $Mappr implements _i1.AutoMapprInterface {
           .toList(),
       itemsWithNullableItem2: model.itemsWithNullableItem2
           .map<_i2.Nested?>(
-              (value) => _map__i2$NestedDto_To__i2$Nested_Nullable(value))
+            (value) => _map__i2$NestedDto_To__i2$Nested_Nullable(value),
+          )
           .toList(),
       tag: null,
     );
@@ -338,8 +324,9 @@ class $Mappr implements _i1.AutoMapprInterface {
     final model = input;
     if (model == null) {
       throw Exception(
-          r'Mapping NestedDto → Nested failed because NestedDto was null, and no default value was provided. '
-          r'Consider setting the whenSourceIsNull parameter on the MapType<NestedDto, Nested> to handle null values during mapping.');
+        r'Mapping NestedDto → Nested failed because NestedDto was null, and no default value was provided. '
+        r'Consider setting the whenSourceIsNull parameter on the MapType<NestedDto, Nested> to handle null values during mapping.',
+      );
     }
     return _i2.Nested(
       id: model.id,
@@ -349,12 +336,14 @@ class $Mappr implements _i1.AutoMapprInterface {
   }
 
   _i2.NestedTag _map__i2$NestedTagDto_To__i2$NestedTag(
-      _i2.NestedTagDto? input) {
+    _i2.NestedTagDto? input,
+  ) {
     final model = input;
     if (model == null) {
       throw Exception(
-          r'Mapping NestedTagDto → NestedTag failed because NestedTagDto was null, and no default value was provided. '
-          r'Consider setting the whenSourceIsNull parameter on the MapType<NestedTagDto, NestedTag> to handle null values during mapping.');
+        r'Mapping NestedTagDto → NestedTag failed because NestedTagDto was null, and no default value was provided. '
+        r'Consider setting the whenSourceIsNull parameter on the MapType<NestedTagDto, NestedTag> to handle null values during mapping.',
+      );
     }
     return _i2.NestedTag();
   }