+++ examples/example/lib/nullable.auto_mappr.dart
@@ -74,20 +74,14 @@ class $Mappr implements _i2.AutoMapprInterface {
   TARGET? tryConvert<SOURCE, TARGET>(
     SOURCE? model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
-      return _safeConvert(
-        model,
-        onMappingError: onMappingError,
-      );
+      return _safeConvert(model, onMappingError: onMappingError);
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvert(
-          model,
-          onMappingError: onMappingError,
-        );
+        return mappr.tryConvert(model, onMappingError: onMappingError);
       }
     }
 
@@ -119,18 +113,16 @@ class $Mappr implements _i2.AutoMapprInterface {
   Iterable<TARGET?> tryConvertIterable<SOURCE, TARGET>(
     Iterable<SOURCE?> model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
       return model.map<TARGET?>(
-          (item) => _safeConvert(item, onMappingError: onMappingError));
+        (item) => _safeConvert(item, onMappingError: onMappingError),
+      );
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertIterable(
-          model,
-          onMappingError: onMappingError,
-        );
+        return mappr.tryConvertIterable(model, onMappingError: onMappingError);
       }
     }
 
@@ -162,7 +154,7 @@ class $Mappr implements _i2.AutoMapprInterface {
   List<TARGET?> tryConvertList<SOURCE, TARGET>(
     Iterable<SOURCE?> model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
       return tryConvertIterable<SOURCE, TARGET>(
@@ -172,10 +164,7 @@ class $Mappr implements _i2.AutoMapprInterface {
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertList(
-          model,
-          onMappingError: onMappingError,
-        );
+        return mappr.tryConvertList(model, onMappingError: onMappingError);
       }
     }
 
@@ -207,7 +196,7 @@ class $Mappr implements _i2.AutoMapprInterface {
   Set<TARGET?> tryConvertSet<SOURCE, TARGET>(
     Iterable<SOURCE?> model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (canConvert<SOURCE, TARGET>(recursive: false)) {
       return tryConvertIterable<SOURCE, TARGET>(
@@ -217,10 +206,7 @@ class $Mappr implements _i2.AutoMapprInterface {
     }
     for (final mappr in _delegates) {
       if (mappr.canConvert<SOURCE, TARGET>()) {
-        return mappr.tryConvertSet(
-          model,
-          onMappingError: onMappingError,
-        );
+        return mappr.tryConvertSet(model, onMappingError: onMappingError);
       }
     }
 
@@ -238,13 +224,8 @@ class $Mappr implements _i2.AutoMapprInterface {
         (targetTypeOf == _typeOf<_i1.User>() ||
             targetTypeOf == _typeOf<_i1.User?>())) {
       if (canReturnNull && model == null) {
-        return (const _i1.User(
-          id: 1,
-          tag: _i1.Nested(
-            id: 1,
-            name: r'default',
-          ),
-        ) as TARGET);
+        return (const _i1.User(id: 1, tag: _i1.Nested(id: 1, name: r'default'))
+            as TARGET);
       }
       return (_map__i1$UserDto_To__i1$User((model as _i1.UserDto?)) as TARGET);
     }
@@ -264,19 +245,13 @@ class $Mappr implements _i2.AutoMapprInterface {
   TARGET? _safeConvert<SOURCE, TARGET>(
     SOURCE? model, {
     void Function(Object error, StackTrace stackTrace, SOURCE? source)?
-        onMappingError,
+    onMappingError,
   }) {
     if (!useSafeMapping<SOURCE, TARGET>()) {
-      return _convert(
-        model,
-        canReturnNull: true,
-      );
+      return _convert(model, canReturnNull: true);
     }
     try {
-      return _convert(
-        model,
-        canReturnNull: true,
-      );
+      return _convert(model, canReturnNull: true);
     } catch (e, s) {
       onMappingError?.call(e, s, model);
       return null;
@@ -293,13 +268,7 @@ class $Mappr implements _i2.AutoMapprInterface {
   _i1.User _map__i1$UserDto_To__i1$User(_i1.UserDto? input) {
     final model = input;
     if (model == null) {
-      return const _i1.User(
-        id: 1,
-        tag: _i1.Nested(
-          id: 1,
-          name: r'default',
-        ),
-      );
+      return const _i1.User(id: 1, tag: _i1.Nested(id: 1, name: r'default'));
     }
     return _i1.User(
       id: model.id,
@@ -314,12 +283,10 @@ class $Mappr implements _i2.AutoMapprInterface {
     final model = input;
     if (model == null) {
       throw Exception(
-          r'Mapping NestedDto → Nested failed because NestedDto was null, and no default value was provided. '
-          r'Consider setting the whenSourceIsNull parameter on the MapType<NestedDto, Nested> to handle null values during mapping.');
+        r'Mapping NestedDto → Nested failed because NestedDto was null, and no default value was provided. '
+        r'Consider setting the whenSourceIsNull parameter on the MapType<NestedDto, Nested> to handle null values during mapping.',
+      );
     }
-    return _i1.Nested(
-      id: model.id,
-      name: model.name,
-    );
+    return _i1.Nested(id: model.id, name: model.name);
   }
 }