+++ packages/auto_mappr/lib/src/builder/value_assignment_builder.dart
@@ -40,10 +40,9 @@ class ValueAssignmentBuilder {
             // Static field.
             ? EmitterHelper.current
                 // ignore: avoid-non-null-assertion, should be ok
-                .refer(sourceField.enclosingElement2.name3!, sourceField.enclosingElement2.library2?.identifier)
+                .refer(sourceField.enclosingElement2.name3!, sourceField.enclosingElement2.library2?.uri.toString())
             // Non static field.
-            // ignore: avoid-nullable-interpolation, should be ok
-            : refer(sourceField.isStatic ? '${sourceField.enclosingElement2.name3}' : 'model'))
+            : refer('model'))
         // ignore: avoid-non-null-assertion, must not be empty
         .property(sourceField.name3!);