+++ packages/auto_mappr/test/integration/enum_mapping_test.dart
@@ -1,7 +1,3 @@
-import 'package:auto_mappr/auto_mappr.dart';
-import 'package:auto_mappr/src/helpers/run_zoned_auto_mappr.dart';
-import 'package:generator_test/generator_test.dart';
-import 'package:source_gen/source_gen.dart';
 import 'package:test/test.dart';
 
 import 'fixture/enum_mapping.dart' as fixture;
@@ -42,53 +38,5 @@ void main() {
     final target = mappr.convert<fixture.EnhancedSourceWithUnknown, fixture.EnhancedTarget>(source);
 
     expect(target, equals(fixture.EnhancedTarget.unknown));
-  });
-
-  group('Error handling', () {
-    test("Can't map source enum to target when target is not enum", () {
-      runZonedAutoMappr(() {
-        final generator = SuccessGenerator.fromBuilder(
-          ['enum_mapping_not_enum.dart'],
-          [],
-          autoMapprBuilder,
-          inputDir: 'test/integration/error_fixture',
-          compareWithFixture: false,
-        );
-
-        expect(
-          generator.test,
-          throwsA(
-            isA<InvalidGenerationSourceError>().having(
-              (x) => x.message,
-              'Match message',
-              'Failed to map Source → Target because target Target is not an enum.',
-            ),
-          ),
-        );
-      });
-    });
-
-    test("Can't map source enum to target when target is not superset of source", () {
-      runZonedAutoMappr(() {
-        final generator = SuccessGenerator.fromBuilder(
-          ['enum_mapping_subset.dart'],
-          [],
-          autoMapprBuilder,
-          inputDir: 'test/integration/error_fixture',
-          compareWithFixture: false,
-        );
-
-        expect(
-          generator.test,
-          throwsA(
-            isA<InvalidGenerationSourceError>().having(
-              (x) => x.message,
-              'Match message',
-              "Can't map enum Source into Target. Target enum is not superset of source enum. (Source → Target)",
-            ),
-          ),
-        );
-      });
-    });
   });
 }